DesVar & costFnc Creation
The DesVar class is functional (creates initial thetavec) and the code is working well using it. Also the call for network has been eliminated from the analysis so it you only call optProb when you run it.
Last but not least, the costFnc class is not working, as it is expected to be used as the gradient and cost gestor.